package renaming --
Anything that was `com.gradle.plugins.*` is renamed to `org.gradle.*`
Don't know why I put com.  its silly.

This is a common plugin library, that holds the schema classes and
slurpers and converters and resolvers.  All of the common library
objects are located at the package of `org.gradle.jacobo.schema.*`.
Notice the lack of a 'plugins' package name.

*.schema.slurper -> *.resolver.  More clear.
*.schema.resolver -> *.resolver.  More clear.
*.schema.document -> *.schema.  superfluous to have the document, when a
 schema is a document

Added packages
*.schema.guice -- holds the guice module

the package org.gradle.jacobo.plugins is there, but only holds three
packages that hold only interface definition that the other plugins will
implement.

*.plugins.ant -- ant executor interface
*.plugins.resolver -- dependency resolver interface
*.plugins.converter -- name converter interface

Build file renames the group to deploy to as org.gradle.jacobo